requirements.txt
pandas
requests
beautifulsoup4
lxml
cssselect
pytz
pyarrow
mypy
pymongo==4.9.2
python-dotenv
pydantic
langchain_openai
langchain_pinecone
langchain
langchain_community